[study] Car 리팩토링 - mvc 패턴으로 리팩토링
