+++ src/main/java/domain/CarRandom.java
@@ -1,0 +1,10 @@
+package domain;
+
+public class CarRandom {
+    private CarRandom() {
+    }
+
+    public static int getRandomNumber() {
+        return (int) (Math.random() * 10);
+    }
+}
