+++ src/main/java/service/RacingCarMain.java
@@ -1,14 +1,14 @@
 package service;
 
 import java.util.List;
-import view.RacingCarView;
+import view.RacingInputCarView;
 
 public class RacingCarMain {
 
     public static void main(String[] args) {
         while (true) {
-            List<String> carNames = RacingCarView.getCarNames();
-            int targetDistance = RacingCarView.getTargetDistance();
+            List<String> carNames = RacingInputCarView.getCarNames();
+            int targetDistance = RacingInputCarView.getTargetDistance();
 
             RacingCarService racingCarService = new RacingCarService();
             racingCarService.startRacing(carNames, targetDistance);
