+++ src/main/java/view/RacingInputCarView.java
@@ -5,7 +5,7 @@ import java.util.List;
 import java.util.Scanner;
 import java.util.stream.Collectors;
 
-public class RacingCarView {
+public class RacingInputCarView {
 
     public static List<String> getCarNames() {
         System.out.println("경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).");
@@ -26,20 +26,4 @@ public class RacingCarView {
         return Integer.parseInt(scan.nextLine());
     }
 
-    public static void printStartRacing() {
-        System.out.println("");
-        System.out.println("실행 결과");
-    }
-
-    public static void printResult(String carNames) {
-        System.out.println("최종 우승자: " + carNames);
-    }
-
-    public static void printNameAndDistance(String name, int distance) {
-        System.out.println(name + " : " + "-".repeat(distance));
-    }
-
-    public static void printNewLine() {
-        System.out.println("");
-    }
 }
